fix keyring-backend=pass
I figured out the monumentally ********* thing that was causing the pass backend to act strange. In onomy there is a Name var that is used throughout to set many things to "onomy". In multiverse I used AppName instead because it was what the ICS people used in their examples. I assumed that this was a constant used by just app.toml to propogate the name around from a single place. It turns out that it actually overrides variables set in version.go in cosmos-sdk. Name is set to an empty string by default. There is also a feature where the build script in the makefile can set AppName via environment variable which is how it set both variables. In the multiverse repo only AppName was set. This would leave it to default to an empty string, and there is another out of the way function that converts this to the default "cosmos" without any warnings. For whatever insane reason and you can ctrl+F version.Name in the SDK, the global Name variable is used by one line that sets the keyring service name, and is used nowhere else, not even in any tests. So nothing except a single issue with one keyring backend manifested any issues. After adding two lines to the makefile it starts working as expected.